tenant refaktoring and tests - part one
