+++ docker-compose.yml
@@ -16,15 +16,5 @@ services:
     ports:
       - "8761:8761"
 
-  dict:
-    image: lapierre/dictionary-service:${DICT_VERSION}
-    environment:
-      - DICT_DATASOURCE_HOST=db:5432
-      - DICT_DATASOURCE_PASSWORD=qwedsazxc
-      - DICT_DATASOURCE_USER=app
-      - DICT_EUREKA_SERVICE_URL=http://eureka:8761/eureka
-    ports:
-      - "9098:9098"
-
 volumes:
   pg_data:
